Update timezone to be of float type (#24)
* Update timezone to be of float type

Reference: https://developers.facebook.com/docs/graph-api/reference/user
Fixes #23

* Update README

+++ profile.go
@@ -2,10 +2,10 @@ package messenger
 
 // Profile is the public information of a Facebook user
 type Profile struct {
-	FirstName     string `json:"first_name"`
-	LastName      string `json:"last_name"`
-	ProfilePicURL string `json:"profile_pic"`
-	Locale        string `json:"locale"`
-	Timezone      int    `json:"timezone"`
-	Gender        string `json:"gender"`
+	FirstName     string  `json:"first_name"`
+	LastName      string  `json:"last_name"`
+	ProfilePicURL string  `json:"profile_pic"`
+	Locale        string  `json:"locale"`
+	Timezone      float64 `json:"timezone"`
+	Gender        string  `json:"gender"`
 }
